image picker bug fix

+++ lib/ui/events/add_event/add_event_viewmodel.dart
@@ -106,22 +106,32 @@ class AddEventViewModel extends BaseViewModel {
 
   void showImagePicker() async {
     setBusy(true);
-    final ImagePicker picker = ImagePicker();
-    image = await picker.pickImage(source: ImageSource.gallery);
-    if (uploadedImageUrl != null) {
-      await s3Service.deleteFile(uploadedImageUrl!['filePath']!);
+    try {
+      final ImagePicker picker = ImagePicker();
+      XFile? temp = image;
+      image = await picker.pickImage(source: ImageSource.gallery);
+      if (image == null) {
+        image = temp;
+        throw 'no image was picked';
+      }
+      if (uploadedImageUrl != null) {
+        await s3Service.deleteFile(uploadedImageUrl!['filePath']!);
+      }
+      double fileSize = double.parse(
+          await HelperFunctions.getFileSize(image!.path, 1, noSuffix: true));
+      if (fileSize <= 6144) {
+        uploadedImageUrl = await s3Service.uploadFile(File(image!.path),
+            s3FolderPath: S3FolderPaths.events);
+        uploaded = true;
+      } else {
+        image = null;
+        //TODO: show image error message
+      }
+      print(await HelperFunctions.getFileSize(image!.path, 1));
+    } catch (e) {
+      print(e);
     }
-    double fileSize = double.parse(
-        await HelperFunctions.getFileSize(image!.path, 1, noSuffix: true));
-    if (fileSize <= 6144) {
-      uploadedImageUrl = await s3Service.uploadFile(File(image!.path),
-          s3FolderPath: S3FolderPaths.events);
-      uploaded = true;
-    } else {
-      image = null;
-      //TODO: show image error message
-    }
-    print(await HelperFunctions.getFileSize(image!.path, 1));
+
     setBusy(false);
     notifyListeners();
   }
